Updated under Amazon S3 data consistency model
in "Buckets have a similar consistency model, with the same caveats. For example, if you delete a bucket and immediately list all buckets, Amazon S3 might still appear in the list."

"Amazon S3 might still appear in the list" --> may either be "the bucket might still appear in the list" or "Amazon S3 might still list the bucket"

Please review.